pyproject.toml
[tool.poetry]
name = "hello-azd-bicep"
version = "0.1.0"
description = "An Azure AI Foundry Agentic AI helloworld."
authors = ["kevinflint-cs2 <kevin.flint@cybersecuritysquared.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
azure-ai-projects = ">=1.0.0,<2.0.0"
azure-identity = ">=1.25.0,<2.0.0"
openai = ">=1.108.0,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.13.1"
mypy = "^1.18.1"
poethepoet = "^0.37.0"
commitizen = "^4.9.1"
detect-secrets = "^1.5.0"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-q"
markers = [
  "smoke: quick smoke tests against Azure AI Foundry",
  "e2e: full end-to-end tests (live Azure calls)"
]

[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "poetry"         # keeps version in pyproject the single source of truth
tag_format = "v$version"
update_changelog = true
changelog_file = "CHANGELOG.md"

[tool.poe.tasks]
fmt = "ruff format ."
fix = "ruff check . --fix"
lint = "ruff check ."
typecheck = "mypy scripts tests"
test = "pytest -q -m smoke"
envsync = 'azd env set --file .env -e dev'
login-az = "az login"
login-azd = "azd login"
deploy = "azd up"
destroy = "azd down --purge --force --no-prompt"
git-add = "git add ."
git-push = "git push -u origin main"
cz-commit = "cz commit"
cleanup-agents = "python ./scripts/cleanup_agents.py"
envs-set = "python ./scripts/set_envs.py"
envs-clean-root = "rm -f .env"
envs-clean-azure = "rm -rf .azure"
secrets-scan = "detect-secrets scan"
secrets-audit = "detect-secrets audit .secrets.baseline"
secrets-validate = "detect-secrets scan --baseline .secrets.baseline"

[tool.poe.tasks.clean]
sequence = ["envs-clean-root", "envs-clean-azure"]

[tool.poe.tasks.checks]
sequence = ["fmt", "fix", "lint", "typecheck", "secrets-validate", "secrets-audit"]

[tool.poe.tasks.push]
sequence = ["git-add", "cz-commit", "git-push"]

[tool.poe.tasks.up]
sequence = ["login-azd", "envs-set", "deploy", "test"]

[tool.poe.tasks.down]
sequence = ["cleanup-agents", "destroy"]